@@ -11,11 +11,12 @@
     "@react-native-community/masked-view": "0.1.10",
     "@react-navigation/material-top-tabs": "^5.3.15",
     "@react-navigation/native": "^5.9.4",
+    "@react-navigation/stack": "^5.14.4",
     "expo": "^38",
     "expo-app-loading": "^1.0.1",
     "expo-cli": "3",
     "expo-status-bar": "~1.0.3",
-    "react": "16.11.0",
+    "react": "^17.0.2",
     "react-dom": "16.11.0",
     "react-native": "0.62.2",
     "react-native-animatable": "^1.3.3",
